FUTURE_CS_03/requirements.txt
Flask
pycryptodome
python-dotenv
